EDMT-302 refactor : mock 폴더 제거 및 notice 도메인 관련 msw 코드 notice 도메인 폴더 하위로 이동
