+++ src/domains/notice/mocks/get-notice-detail.ts
